Fix Windows compilation
