+++ pkg/useragent/uname_unix_test.go
@@ -1,0 +1,14 @@
+// +build !windows
+
+package useragent
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestUnameNotEmpty(t *testing.T) {
+	u := getUname()
+	require.NotEmpty(t, u)
+}
